+++ src/main/java/com/mishkapp/minecraft/plugins/squarekit/suffixes/stats/holding/HoldingSpeed.java
@@ -14,17 +14,17 @@ import java.util.HashMap;
  * Created by mishkapp on 03.10.2016.
  */
 public class HoldingSpeed extends Suffix {
-    private float speed;
+    private double speed;
 
     public HoldingSpeed(KitPlayer kitPlayer, ItemStack itemStack, Integer level) {
         super(kitPlayer, itemStack, level);
         if(level > 32){
-            speed = (float) (-1 * (level - 31) * (100.0/32));
+            speed = -1 * (level - 31) * (100.0/32);
         } else {
-            speed = (float) ((100.0/31) * level);
+            speed = (100.0/31) * level;
         }
 
-        speed /= 100;
+        speed /= 100.0;
     }
 
     @Override
@@ -33,11 +33,11 @@ public class HoldingSpeed extends Suffix {
     @Override
     public void handle(KitEvent event) {
         if(event instanceof SuffixTickEvent){
-            HashMap<Suffix, Float> adds = kitPlayer.getSpeedAdds();
+            HashMap<Suffix, Double> adds = kitPlayer.getSpeedAdds();
             if(isItemHolding()){
                 adds.put(this, speed);
             } else {
-                adds.put(this, 0.0f);
+                adds.put(this, 0.0);
             }
         }
         kitPlayer.updateStats();
